more typescript improvements
