+++ src/shared/types/index.ts
@@ -122,95 +122,3 @@ export interface AvailableAccount {
   photoCount: number;
   feedCount: number;
 }
-
-export interface ElectronAPI {
-  collectPhotos: (params: {
-    accountId: string;
-    token?: string;
-    forceAccountsRefresh?: boolean;
-    forceRoomsRefresh?: boolean;
-  }) => Promise<ApiResponse<CollectionResult>>;
-
-  collectFeedPhotos: (params: {
-    accountId: string;
-    token?: string;
-    incremental?: boolean;
-    forceAccountsRefresh?: boolean;
-    forceRoomsRefresh?: boolean;
-  }) => Promise<ApiResponse<CollectionResult>>;
-
-  downloadPhotos: (params: {
-    accountId: string;
-  }) => Promise<ApiResponse<DownloadResult>>;
-
-  downloadFeedPhotos: (params: {
-    accountId: string;
-  }) => Promise<ApiResponse<DownloadResult>>;
-
-  selectOutputFolder: () => Promise<string | null>;
-  getSettings: () => Promise<RecNetSettings>;
-  updateSettings: (
-    settings: Partial<RecNetSettings>
-  ) => Promise<RecNetSettings>;
-  getProgress: () => Promise<Progress>;
-  cancelOperation: () => Promise<boolean>;
-  onProgress: (callback: (event: any, progress: Progress) => void) => void;
-  removeProgressListener: (
-    callback: (event: any, progress: Progress) => void
-  ) => void;
-  lookupAccount: (accountId: string) => Promise<ApiResponse<AccountInfo[]>>;
-  searchAccounts: (username: string) => Promise<ApiResponse<AccountInfo[]>>;
-  clearAccountData: (
-    accountId: string
-  ) => Promise<ApiResponse<{ filesRemoved: number }>>;
-  loadPhotos: (accountId: string) => Promise<ApiResponse<Photo[]>>;
-  loadFeedPhotos: (accountId: string) => Promise<ApiResponse<Photo[]>>;
-  listAvailableAccounts: () => Promise<ApiResponse<AvailableAccount[]>>;
-  loadAccountsData: (accountId: string) => Promise<ApiResponse<PlayerResult[]>>;
-  loadRoomsData: (accountId: string) => Promise<ApiResponse<RoomDto[]>>;
-  loadEventsData: (accountId: string) => Promise<ApiResponse<EventDto[]>>;
-
-  // Favorites management
-  getFavorites: () => Promise<ApiResponse<string[]>>;
-  toggleFavorite: (photoId: string) => Promise<ApiResponse<boolean>>;
-  isFavorite: (photoId: string) => Promise<ApiResponse<boolean>>;
-
-  // Open external URL in system browser
-  openExternal: (url: string) => Promise<void>;
-
-  // Auto-updater
-  checkForUpdates: () => Promise<void>;
-  downloadUpdate: () => Promise<void>;
-  installUpdate: () => Promise<void>;
-  getAppVersion: () => Promise<string>;
-  onUpdateAvailable: (
-    callback: (info: {
-      version: string;
-      releaseDate?: string;
-      releaseNotes?: string;
-    }) => void
-  ) => void;
-  onUpdateNotAvailable: (callback: () => void) => void;
-  onUpdateDownloadProgress: (
-    callback: (progress: {
-      percent: number;
-      transferred: number;
-      total: number;
-    }) => void
-  ) => void;
-  onUpdateDownloaded: (callback: (info: { version: string }) => void) => void;
-  onUpdateError: (callback: (error: { message: string }) => void) => void;
-  removeUpdateListeners: () => void;
-
-  // Window controls
-  windowMinimize: () => Promise<void>;
-  windowMaximize: () => Promise<void>;
-  windowClose: () => Promise<void>;
-  windowIsMaximized: () => Promise<boolean>;
-}
-
-declare global {
-  interface Window {
-    electronAPI: ElectronAPI;
-  }
-}
